requirements.txt



python-dotenv>=1.0.0  # For loading .env file

# Voice-to-Text (API-based or local):
requests>=2.31.0  # For HuggingFace Inference API
openai-whisper>=20231117  # Optional - for local Whisper transcription (fallback if API fails)

# LLM Support (API-based, no local models):
groq>=0.4.0  # For Groq API (Llama models) - Recommended


sounddevice>=0.4.6  # Optional - only needed for microphone input
soundfile>=0.12.1   # Optional - only needed for microphone input


fastapi>=0.104.0
uvicorn[standard]>=0.24.0  # ASGI server for FastAPI
python-multipart>=0.0.6  # Required for FastAPI file uploads

